fix: correct @tf.function application for tf.keras 2.2
In tf.keras 2.2, for models that subclass `tf.keras.Model` eager execution is enabled
but for models that do not subclass, eager execution is not enabled during `_aggregate_gradients`
and `apply_gradients`. We can not apply @tf.function in 2.2, since we do not know if the users
model subclasses `tf.keras.Model`.

To support running in non-eager mode without @tf.function we have to convert some our graph
operations to be python operations.